requirements.txt
credentialsmanager @ file:///home/fewill/code/credentialsmanager
onepassword-sdk
pyyaml>=6.0
python-dotenv
slack-sdk
boto3
